Update os_type.rs
I want to use u8 to record what system it is, enum can be as u8 . Unknown is placed first, and it will be the number 0. I think it is appropriate to use 0 to identify unknown systems.
In addition, I hope that when additional operating systems are added in the future, the ordering of the existing systems in the enum will not be changed.